[Java] Remove `AbstractMessageTransceiver`
